requirements.txt
flask
groq
langchain
langchain_community
langchain_openai
pypdf
streamlit
typing
python-dotenv
